Refactor entity tests.
