+++ server/src/test/java/org/spine3/server/entity/EntityEqualsShould.java
@@ -23,6 +23,8 @@ package org.spine3.server.entity;
 import com.google.protobuf.Timestamp;
 import org.junit.Before;
 import org.junit.Test;
+import org.spine3.test.entity.Project;
+import org.spine3.testdata.Sample;
 
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotEquals;
@@ -41,17 +43,16 @@ import static org.spine3.test.Tests.nullRef;
  */
 public class EntityEqualsShould {
 
-    //TODO:2017-02-01:alexander.yevsyukov: Have local test class instead of Given.TestEntity.
-    private Given.TestEntity entity;
+    private TestEntity entity;
 
     @Before
     public void setUp() {
-        entity = Given.TestEntity.withState();
+        entity = TestEntity.withState();
     }
 
     @Test
     public void assure_same_entities_are_equal() {
-        final Given.TestEntity another = Given.TestEntity.withState(entity);
+        final TestEntity another = TestEntity.withState(entity);
 
         assertTrue(entity.equals(another));
     }
@@ -75,7 +76,7 @@ public class EntityEqualsShould {
 
     @Test
     public void assure_entities_with_different_ids_are_not_equal() {
-        final Given.TestEntity another = Given.TestEntity.newInstance(newUuid());
+        final TestEntity another = TestEntity.newInstance(newUuid());
 
         assertNotEquals(entity.getId(), another.getId());
         assertFalse(entity.equals(another));
@@ -83,8 +84,8 @@ public class EntityEqualsShould {
 
     @Test
     public void assure_entities_with_different_states_are_not_equal() {
-        final Given.TestEntity another = Given.TestEntity.withState(entity);
-        another.setState(Given.newProject(), another.getVersion(), another.whenModified());
+        final TestEntity another = TestEntity.withState(entity);
+        another.setState(Sample.messageOfType(Project.class), another.getVersion(), another.whenModified());
 
         assertNotEquals(entity.getState(), another.getState());
         assertFalse(entity.equals(another));
@@ -92,7 +93,7 @@ public class EntityEqualsShould {
 
     @Test
     public void assure_entities_with_different_versions_are_not_equal() {
-        final Given.TestEntity another = Given.TestEntity.withState(entity);
+        final TestEntity another = TestEntity.withState(entity);
         another.setVersion(entity.getVersion() + 5, another.whenModified());
 
         assertFalse(entity.equals(another));
@@ -100,7 +101,7 @@ public class EntityEqualsShould {
 
     @Test
     public void assure_entities_with_different_modification_times_are_not_equal() {
-        final Given.TestEntity another = Given.TestEntity.withState(entity);
+        final TestEntity another = TestEntity.withState(entity);
         another.setVersion(another.getVersion(), Timestamp.newBuilder().setSeconds(5).build());
 
         assertNotEquals(entity.whenModified(), another.whenModified());
